+++ models/transaction.go
@@ -6,10 +6,10 @@ import "gorm.io/gorm"
 
 type Transaction struct {
 	gorm.Model
-	Amount		uint		`json:"amount"`
+	Amount		int64		`json:"amount"`
 	Type 			string	`json:"type"`
 	Reference		string	`json:"reference"`
 	Status		string	`json:"status" gorm:"default:pending"`
-	User_id		int		`json:"user_id"`
+	User_id		uint		`json:"user_id"`
 	User			User
 }
